Refactor compiler to eliminate `depth`
- The notion of `depth` within the compiler was a little convoluted; in
addition to it delineating which function scope (wrt nested functions) a
variable was defined in, it also described loops and ifs. This is
especially useful when determining which locals to pop when the scope
ends.
- The refactoring here involved removing as much logic as possible that
was reliant on `depth` and instead use an abstraction of a Scope. This
abstraction is _not_ coupled with a notion of `depth`. The notion of
`depth` will be used instead to handle upvalues and closures.