Cleanup: cleaning up TypedAstNode::Break
- Since the `loop_depth` field was no longer being used in the compiler
(we have a separate mechanism for tracking this type of thing now, as
of #85), it can be removed from the ast node.
- Also add logic (and tests) to verify that a break placed anywhere
other than within a Loop scope is invalid (a loop placed in a Block
scope will be okay, as long as it eventually is contained by a Loop).