Add VM support for closures + upvalues
There are a _ton_ of changes here, so I'll try and summarize by area of
concern.
- Typechecker:
  - The `TypedFunctionDeclNode` node now has an `is_recursive` field.
This is necessary in order to facilitate certain compilation strategies
for nested recursive functions.
  - There's also a bit of a cleanup with how `NativeFn`s are
typechecked; see the NativeFns summary for more details

- Compiler:
  - Fix issue where Upvalues which reference an upvalue_idx could not
themselves reference another such Upvalue. Previously, all that had
worked was Upvalues which reference an upvalue_idx could only point to
an Upvalue which referenced a local_idx.
  - Use the `is_recursive` field when compiling nested functions to save
an initial dummy value (the value 0) to a local slot, and then
overwriting it with the actual function value once its compilation is
complete. This allows for functions to be captured within themselves as
upvalues, thus facilitating recursive nested functions.

- VM:
  - Adding runtime support for capturing and closing local variables
into upvalues, as well as upgrading functions into closures if they
close over any variables.
  - Also added the `Value::Closure` type. This could stand to be
refactored a bit since it's so similar to `Value::Fn`, but that's for
another day.

- NativeFns
  - This was a bit of an unexpected change, but it _was_ necessary.
Because of the implementation of `Value::Closure` and how it needed
to have shared, mutable, heap-allocated data to represent its upvalues,
there were issues initializing the `NATIVE_FNS` and `NATIVE_FNS_MAP`
using `lazy_static!`. I changed the signature of `NativeFn` a bit to
still achieve the notion of optional values though, and I actually think
this design is much nicer. But man, did it take a while to figure out
what was the underlying issue there. I still have a lot to learn about
how rust deals with shared, mutable, heap-allocated memory.