Changing default-valued param funcs
- Adding TypedAstNode::_Nil, which should never be directly parsed, but
is used when dynamically generating the AST for default param value
evaluation bytecode.
- Remove the old (and admittedly pretty janky) way of handling functions
with default param values. Now, all functions are invoked with the same
number of args as its arity, with `Nil` values being placed on the stack
when no value was passed for that argument. Within the function body,
bytecode is inserted to reassign to the param's local if `Nil` was
passed to it.