Change implementation of Coalesce
- Remove `Obj::OptionObj`. The notion of an "Option<T>" needs to only
exist at typechecking-time, and doesn't require a runtime box type. At
runtime, an instance of `Option<T>` will either be `Nil` or an instance
of `T`, and the Coalesce operator `?:` operates on `Nil`s now. This
reduces some overhead with the `OptMk` opcode, as well as simlifying the
Object model (removing `OptionObj`).