Fixing newline parsing issues
- There had been an issue where newlines were being ignored in the
middle of any binary expression. This resulted in code such as
`println("hello")\n[1]` to be treated as an array index operation,
rather than a top-level statement and a top-level expression. Correctly
recognizing newlines within `parse_precedence` fixes this issue.
However, it's important that we only break when the _next_ expression to
be parsed is of `Precendence::Call` and we had just seen a `Newline`; if
we don't do this, then expressions such as `1\n+\n1` would be
impossible, which doesn't really make sense.
- This required a change to the lexer, which now can emit a `Newline`
token. This token will represent any block of newlines; there should
never be more than 1 consecutive `Newline` token in a token stream.