Fixing newline parsing issues
- There had been an issue where newlines were being ignored in the
middle of any binary expression. This resulted in code such as
`println("hello")\n[1]` to be treated as an array index operation,
rather than a top-level statement and a top-level expression.
- To fix this, first I had tried a more complex approach where I tracked
`Token::Newline`s, a new type of Token, and tried to make decisions at
parse-time. This proved to be difficult to do, due to the way that the
Pratt parsing technique attempts to aggregate expressions (see #107 for
the failed attempt). The better way to fix this was to modify the
`Token::LBrack` token to know whether it was preceded by a newline
character, and if so, treat it as `Precedence::None` in the parser (so
it would be treated as an array literal); otherwise, it'd retain its
normal `Precedence::Call` (so it'd be treated as an index operator).
- This is the only place where this needs to happen (at the moment),
since the `[` token is the only token that can be both an infix operator
and the start of a more complex expression. I expect that if tuples were
to exist (something like `(1, 2)`), a similar approach would need to be
taken in order to avoid errant `Precedence::Call` parses.