Smoothing out the definition of native methods
- Split out the data needed for typechecking from the actual data needed
at runtime. This helps ensure that no extra data is passed around when
it doesn't need to be, and it also makes the implementation cleaner.
Still looking for a way to maybe have a macro which cleans up the
implementations of `NativeType` for `NativeString` and `NativeArray`.
But that can come later.