Fixing parsing bug
- Right now, parsing a grouped expression preceded by an array literal
declaration fails because it treats the `Token::LParen` of the
grouped expression as an invocation. This is similar to an issue I've
resolved in the past where a `Token::LBrack` after an expression was
being parsed as an indexing operation. The solution here is the same as
it was then (until I can think of something better) - if the token in
question (here the `Token::LParen`) is preceded by a newline, don't
parse it with `Precedence::Call`. This means that the following
code will be successfully parsed as intended:
```
val nums = [1, 2, 3]
(1 + 2)
```