Optional-safe accessor
Addresses #124

- Add lexing of `?.` operator
- Add parsing of `?.` expressions (it creates an `AstNode::Accessor`
whose `AccessorNode` has `is_opt_safe: true`).
- Add typechecking of `?.` expressions. `a?.b` will "unwrap" `a` if it's
an optional, and then typechecks the field `b` against it, then re-wraps
it as an optional; if `a` is _not_ optional, it's effectively the same
as `a.b`; a `Token::Dot` is actually emitted instead of
`Token::QuestionDot`, so downstream code will actually compile away the
unnecessary optional-safeness.