Optional-safe accessors: compilation
- This is some interesting code. I'm not sure how stable it'll be, but
it seems to work for now. Basically when evaluating a chain of accessor
expressions that includes an opt-safe one `?.`, unwind the chain and
rewrite it as a series of nested if-exprs.
- Note also that the type equivalence rules of assigning `A` to `A?`
were changed too. This should probably be pulled out into a separate
issue, but I needed to resolve this in order to have a sensible example
to work off of.