Bugfix: Using unclosed upvalues
- There was a bug where accessing upvalues in a function that have not
yet been closed over breaks. Normally, when we close over a local in a
function by another function, then return that function we're okay since
that closed-over value is moved onto the heap by the time the returned
closure executes. But if we call the closure before the enclosing
function returns, we never close over that captured local.
- In the code to load and store upvalues within the VM, there was a bug.
If the upvalue was not yet closed over (as is the case here), we were
using the upvalue's `slot_idx` and loading/storing a local, _but we were
doing it with respect to the current frame_. Instead, we want to do it
against the base of the stack in general, since this value is being
evaluated with respect to the caller's frame, and not the callee's
frame. What a sneaky nefarious little bug