Proper if-exprs
- I realized that if-expressions don't actually work properly... They're
fine if there's nothing else on the stack, but if there is, then any
variables created in the meantime mess things up. For example
```
var a = 1
1 + if (true) {
  val b = 2
  a
} else { 0 }
```
The variable `b` means that the binding `a` will not be properly
resolved, since the `1` will pollute the stack in the meantime.
- The "solution" to this (still not sure how I feel about it) is to have
if-expressions compile down to invocations of anonymous functions, the
bodies of which consist solely of the if-expression. This properly
resolves any stack issues (since variables referenced from outside of
the if-expr will be via upvalues). What I haven't yet decided though is
whether this convenience is _always_ worth allocating an anonymous
function for. For example, if there is nothing on the stack (which we
should be able to know at compile-time) we should be smart enough to not
incur this cost:
```
val a = if (true) { 1 } else { 2 }
```
This should ideally not create and immediately invoke an anonymous
function. But that sounds like a future enhancement, right?
- Also of note, since optional-safe accessors compile to if-expressions
under the hood, that code also needed to be wrapped in an
immediately-invoked function expression. Now, expressions like
```
1 + (maybeString?.length ?: 0)
```
will work, when previously it caused all bunch of stack-related issues.