Condition convenience bindings: Typechecking
- Add typechecking for condition convenience bindings in if-stmts/exprs
and while-loops. This binding will either be a Boolean if the type of
the condition is Boolean (in fact, it will always be `true`); if the
condition is of type `T?`, the binding will be of type `T`. This will
also unwind Optionals, so if the condition is of type `T???`, the
binding will be of type `T`.