Cond convenience bindings (#148)
* Condition convenience bindings

- Adding lexing for the pipe (`|`) token

* Conditional convenience bindings

- Parsing of conditional bindings for if-stmts/exprs or while-loops.
This also fixes a lexing bug where the first character after the Pipe
token was skipped over.

* Condition convenience bindings: Typechecking

- Add typechecking for condition convenience bindings in if-stmts/exprs
and while-loops. This binding will either be a Boolean if the type of
the condition is Boolean (in fact, it will always be `true`); if the
condition is of type `T?`, the binding will be of type `T`. This will
also unwind Optionals, so if the condition is of type `T???`, the
binding will be of type `T`.

* Condition convenience bindings: Compilation

- Compilation of convenience bindings for if-stmts/exprs and
while-loops. This injects a new binding into the block which is a `Dup`
of the condition value on the stack.