Rewrite coalescence in terms of if-expr (#154)
Contributes to #153

- Within the compiler, we compile a coalesce operation (`?:`) as an
if-expression with a condition binding. This unifies the bytecode
generated for both of them, and helps keep all the jumping logic in one
place.
- Note that this also fixed a bug where locals created in the thin gap
between an if-expr's condition being executed and either its then or
else block could potentially remain on the stack. If there is a
condition binding (which gets duplicated and placed on the stack), and
we do _not_ enter the `then` branch, that value was left to hang out on
the stack. We need to make sure that in the else branch, that value is
`Pop`ped properly.