Rewrite Coalesce and Opt-safe Accessors
- Operations involving the Coalesce operator (`?:`) and Option-safe
Accessors (`?.`) are now rewritten to use if-expressions and condition
bindings. This also wraps those expressions (as well as if-expressions
themselves) in IIFEs (Immediately-Invoked Function Expressions) in order
to prevent pollution of the calling scope.
- Note that "function expressions" do not yet exist as a language
feature, they're mostly a hack at the moment (`is_anon` on `TypedFunctionDeclNode`).
But that'll change in a coming PR.