Reference-type indirection layer (#156)
- When building up complex types (ie. struct type declarations), it
becomes increasingly important to handle cycles in type data. Meaning,
if a type `Foo` had a method `foo` which returned `Foo`, we have a "cycle",
and the representation of that Type as a data structure becomes
difficult to work with. For example, at typecheck-time, the methods of
the return value of `Foo#foo()` will be unknown, and saying
```
val f = Foo()
f.foo().foo()
```
will not properly typecheck.
- Add an indirection layer of `Type::Reference`, which wraps a `String`
which serves as a key into a hashmap stored within the typechecker. That
hashmap's value for a given type name will be the real structure of the
type, which is subject to change and grow outside of its References.
  - Note that this is not a memory-enforced reference (like a pointer or
memory-reference); this is a loose agreement that any `Type::Reference`
value will have a value present in the typechecker's
`referencable_types` hashmap. Kind of a loose abstraction, but not
leaky.