Lambdas: Typechecking
- Add typechecking for lambda expressions
- Change the way types are tested for equivalence, since we need to
perform inference-lite, in order account for lambdas with untyped
parameters. (We _do_ forbid binding declaration to have an instance of
`Type::Unknown` anywhere within its type.) During the typechecking
phase, we _can_ mutate the underlying type of the node being tested for
equivalence, if it's a lambda node with an arg of type `Type::Unknown`.
- I imagine future work related to inference (if/when I ever get around
to doing that) will clean this up a bit?