Compilation of lambdas (#163)
- Compile TypedAstNode::Lambda nodes!
- Modifies the LStore instruction to push a new value if the stack slot
does not exist _but_ it would be the next available slot. This is
essential to making default-valued arguments work in lambdas where the
type is not guaranteed to be known ahead of time:
```
func call(fn: (Int) => Int, value: Int) = fn(value)
call((x, y = 1, z = 2) => x + y + z, 3) // returns 6
```
- Note that this does _not_ cover inference in cases like:
```
((a, b) => a + b)(1, 2)

val fn: ((Int) => Int)? = [
  x => x + 1,
  y => y + 1
][0]
```
All of this will need to be addressed, along with less esoteric
examples, such as
```
val nums: Int[] = []
```
which is currently being handled in a one-off special case.