Lambda Expressions (#157) (#158)
* Lambda Expressions (#157)

- Add parsing of lambda expressions. This supports no-arguments,
single-arguments (with and without parens), and multi-argument lambdas.
- This also refactors the funcdecl parsing, since it pulls out the
argument parameter parsing.
- Note: This pulls in the `peekmore` crate to peek additional tokens
ahead in the stream. The stdlib Peekable trait only allows for one
lookahead item.

* Lambdas: Typechecking (#160)

* Lambdas: Typechecking

- Add typechecking for lambda expressions
- Change the way types are tested for equivalence, since we need to
perform inference-lite, in order account for lambdas with untyped
parameters. (We _do_ forbid binding declaration to have an instance of
`Type::Unknown` anywhere within its type.) During the typechecking
phase, we _can_ mutate the underlying type of the node being tested for
equivalence, if it's a lambda node with an arg of type `Type::Unknown`.
- I imagine future work related to inference (if/when I ever get around
to doing that) will clean this up a bit?

* Lambdas: Type Idents (#159)

- This adds type identifier parsing logic for function types; function
types are of the form `(A, B, C) => D`.

* Lambdas: Adding Unit type (#161)

- We need to add a `Unit` type that can be referenced in function type
identifiers, since otherwise there's no way to represent the type of
this function:
```
val fn = a => println(a)
```
- This also fills in some missing implementations for TypecheckerErrors

* Compilation of lambdas (#163)

- Compile TypedAstNode::Lambda nodes!
- Modifies the LStore instruction to push a new value if the stack slot
does not exist _but_ it would be the next available slot. This is
essential to making default-valued arguments work in lambdas where the
type is not guaranteed to be known ahead of time:
```
func call(fn: (Int) => Int, value: Int) = fn(value)
call((x, y = 1, z = 2) => x + y + z, 3) // returns 6
```
- Note that this does _not_ cover inference in cases like:
```
((a, b) => a + b)(1, 2)

val fn: ((Int) => Int)? = [
  x => x + 1,
  y => y + 1
][0]
```
All of this will need to be addressed, along with less esoteric
examples, such as
```
val nums: Int[] = []
```
which is currently being handled in a one-off special case.

* Some minor lambda cleanups (#164)

- Replace the FunctionDecl in the IIFE implementation with Lambda, and
remove the `is_anon` field from FunctionDecl.
- Fix issue where values closed over by a lambda in any block other than
a function were not being properly captured; we need to emit a
`CloseUpvalueAndPop` instruction in this case.