Marking locals (#167)
- Relying on stack location as a 1-1 mapping to local index is somewhat
problematic, particularly when you need to resolve locals when the stack
may be in a polluted state mid-expression. A great example of this is:
```
{
  val a = 1
  a
} + {
  val b = 2
  b
}
```
Resolving the value for `b` (which is local index 0 in this case) would
face interference from the value for `a` (1) which is still on the
stack. This means that we need some other way of mapping locals to stack
addresses, so we introduced the `local_addr` local lookup table (just a
vector, since the keys are monotonically increasing). When resolving a
local by its index (during loads, stores, upvalue resolution), we look
to this table. Values in this table are populated via the
newly-introduced `MarkLocal <local_idx>` instruction, which lets the VM
know that the value currently on top of the stack should be treated as a
local, and thus stored in the `local_addr` table for the current frame.
- Dealing with functions and their arguments is a little bit annoying
here, especially when it comes to lambdas and default-valued arguments.
Function parameters (including the `<ret>` reserved slot) are placed
onto the stack _before_ the function's code executes, so calling
`MarkLocal` on those values would happen too late. Instead, what happens
is the handling of the `Invoke` instruction will pre-fill those values
as it creates the `CallFrame`. Lambdas are annoying because if they have
arguments with default values, those values would not be pushed onto
the stack before the function call, since the arity of the function is
dictated by the caller, not the callee. The nil-checking code still runs
within the function body, but when it does its `LStore` operation, there
will not be a value in the frame's `local_addrs` table, since it wasn't
pre-filled when the `CallFrame` was created, and there was no
`MarkLocal` instruction for it. To resolve this, if there is no value
yet in `local_addrs`, we just push the topmost stack slot. This is
likely problematic and probably going to come back to bite me, but it
works for now so....
- Also, refactor some of the functions in VM to be colocated with relevant
functions, and just some other general cleanup.