Fixing Bug: conditional binding with no else-block
- There was a bug I stumbled across when playing around with writing
code to test generics (which had nothing to do with generics). If we had
an if-statement (a _statement_, not an expression) with a
conditional-binding and _no else block_, we'd end up double-emitting the
`POP` for that conditional binding - once within the if-block, and once
before the else-block (even though the else-block didn't exist).
Example:
```
if true |a| {
  println("hello")
}
```
This will trigger a StackEmpty runtime error due to the double-`POP`. I
fixed this by only emitting that `POP` within the else-block (if a
conditional binding was present), and compiling a fake dummy empty
else-block if none was provided in source. We can always assume there'll
be an else-block if a conditional-binding is present.