Generics: Adding native array methods
- Adding `map`/`filter`/`reduce` methods for Array:
  - This requires proper handling of generics for NativeFns
  - This also requires some way of invoking bytecode-backed
functions/lambdas from native rust-land, which required some refactoring
within the VM. As a result of this, NativeAbraFns now accept a mutable
reference to the VM. Also the mechanisms surrounding function invocation
needed to be changed a bit.
- Changes to function invocation:
  - The `Invoke` opcode no longer accepts 2 arguments, instead the
`has_return` is obtained from the FnValue/ClosureValue/NativeFnValue
itself. This was necessary in order to programmatically invoke
bytecode-backed function values from rust-land.