Generics: struct types (#175)
There's a ridiculous amount of changes here, let's go module-by-module:

Parser
- Parsing type identifiers with type arguments `List<Int>`

Types
- Add a list of type parameters to `Reference` types
- Add comparison logic to `is_equivalent_to` for comparing struct types'
type arguments. A note here: if the given type of a struct type arg
comparison is a `Generic`, we give it a pass.
- Cleanup some of the reference type stuff into the `hydrate_reference_type`
helper function
- Did some renaming of "generics" in functions to `unbound_generics` to better
communicate intent
- Adding functionality to `from_type_ident` to return type identifiers with type
arguemnts as `Reference`s with type_args.

Typechecker:
- Add `type_from_type_ident` for some cleanup, as well as returning an
error when the arity of type arguments doesn't match
- Resolving generics with struct types, for fields and methods
- Needed to do some cleanup in a couple of places related to
`self.scopes.push`/`self.scopes.pop` in order to make sure that the
generic type args were within scope at the time of typechecking.
- A type's fields should be able to be referenced within its methods;
I should be able to construct a new instance of a type within a method.