Generics (#176)
* Generics: Parsing (#171)

- Parsing type arguments for generic type definitions

* Generics: Typechecking bare functions (#174)

* Generics: Typechecking bare functions

- `Type::Fn(_, _)` => `Type::Fn(FnType { .. })`
- Adding type arguments to functions
- Resolve generics on function invocation, for both named and unnamed
argument calling convention

* Fixing Bug: conditional binding with no else-block

- There was a bug I stumbled across when playing around with writing
code to test generics (which had nothing to do with generics). If we had
an if-statement (a _statement_, not an expression) with a
conditional-binding and _no else block_, we'd end up double-emitting the
`POP` for that conditional binding - once within the if-block, and once
before the else-block (even though the else-block didn't exist).
Example:
```
if true |a| {
  println("hello")
}
```
This will trigger a StackEmpty runtime error due to the double-`POP`. I
fixed this by only emitting that `POP` within the else-block (if a
conditional binding was present), and compiling a fake dummy empty
else-block if none was provided in source. We can always assume there'll
be an else-block if a conditional-binding is present.

* Generics: Adding native array methods (#173)

- Adding `map`/`filter`/`reduce` methods for Array:
  - This requires proper handling of generics for NativeFns
  - This also requires some way of invoking bytecode-backed
functions/lambdas from native rust-land, which required some refactoring
within the VM. As a result of this, NativeAbraFns now accept a mutable
reference to the VM. Also the mechanisms surrounding function invocation
needed to be changed a bit.
- Changes to function invocation:
  - The `Invoke` opcode no longer accepts 2 arguments, instead the
`has_return` is obtained from the FnValue/ClosureValue/NativeFnValue
itself. This was necessary in order to programmatically invoke
bytecode-backed function values from rust-land.

* Generics: struct types (#175)

There's a ridiculous amount of changes here, let's go module-by-module:

Parser
- Parsing type identifiers with type arguments `List<Int>`

Types
- Add a list of type parameters to `Reference` types
- Add comparison logic to `is_equivalent_to` for comparing struct types'
type arguments. A note here: if the given type of a struct type arg
comparison is a `Generic`, we give it a pass.
- Cleanup some of the reference type stuff into the `hydrate_reference_type`
helper function
- Did some renaming of "generics" in functions to `unbound_generics` to better
communicate intent
- Adding functionality to `from_type_ident` to return type identifiers with type
arguemnts as `Reference`s with type_args.

Typechecker:
- Add `type_from_type_ident` for some cleanup, as well as returning an
error when the arity of type arguments doesn't match
- Resolving generics with struct types, for fields and methods
- Needed to do some cleanup in a couple of places related to
`self.scopes.push`/`self.scopes.pop` in order to make sure that the
generic type args were within scope at the time of typechecking.
- A type's fields should be able to be referenced within its methods;
I should be able to construct a new instance of a type within a method.

* Generics: fixing struct type issue (#177)

- When a type ident is initially assigned its type, it is assigned a
Reference type with an empty type_args vector. Once we are aware of its
type_args, we need to update the binding's type to have generic
placeholders, so that downstream lookups can do proper generic
substitutions.

* Generics: Methods (#178)

- Change to parser: instances of `AstNode::Identifier` can now have type
arguments, if the next character is a `(` (denoting a function call with
provided type arguments).
- Changes to typechecker: when method fields are accessed off of
bindings, make sure to perform any generic substitions when determining
the type of that field, if generics are provided and the function itself
can accept type arguments.