Addresses #188
- We need to handle deep method call stacks which start in functions
called by native functions. For example:
```
[1, 2].map(i => someFunction(i))
```
If `someFunction` calls several other functions, we need to make sure
the stack stays clean. As it was right now, the `run` method had no way
of distinguishing between the return from a function deep in the call
stack and a return of `someFunction`. So, when calling `run` to execute
the bytecode of a function triggered from a native context, make sure to
specify the frame depth of the call site.