Enums (#190)
* Enums: Lexing (#184)

* Enums  parsing (#185)

* Enums: Lexing

* Enums: Parsing

* Add support for enum variants with arguments

* Enums: Typechecking (#186)

* Enums: Typechecking

- Typechecking declaration of basic variants _and_ constructor variants,
as well as usages of them

* Exclude 'self' as a valid param for a constructor enum variant, it throws an error now

* Typechecking enum methods and static methods

* Enums: Compilation (#187)

- Typechecking of static methods on enums
- Compilation of enums! And tests

* Fixing native fn bug (#189)

* Enums: Compilation

- Typechecking of static methods on enums
- Compilation of enums! And tests

* Addresses #188

- We need to handle deep method call stacks which start in functions
called by native functions. For example:
```
[1, 2].map(i => someFunction(i))
```
If `someFunction` calls several other functions, we need to make sure
the stack stays clean. As it was right now, the `run` method had no way
of distinguishing between the return from a function deep in the call
stack and a return of `someFunction`. So, when calling `run` to execute
the bytecode of a function triggered from a native context, make sure to
specify the frame depth of the call site.

* Removing unnecessary 'continue'