Match statements/expressions: handling Type cases
- Match statements and expressions can now match a value against a given
type! So the things that can be matched against now include `None`,
types, and a wildcard (`_`).
- This also refactors `Prelude` to use `thread_local!` for static global
constants, to help tidy things up a bit.
- This also adds a way to compare types for equality (as is necessary in
order to facilitate Type branches in `match` statements/expressions);
right now it just compares the name, but it should probably be more
robust than that in the future.