Adding enum variants as separate types
- Add new type `Type::EnumVariant`, which is a nicer abstraction over
top of enum variants, allowing each enum's variant to represent a
separate discrete type. This allows us to access fields of an enum
variant, ie
```
enum Foo { Bar(baz: Int) }
val f = Foo.Bar(baz: 24)
f.baz
```
Since the typechecker now knows that the type of `f` is not only `Foo`,
but `Foo.Bar`, it allows us to access `.baz` off of it. In prior
implementations, this would have been typechecked as:
```
val f: Foo = Foo.Bar(baz: 24)
f.baz
 ^ Error, no member called 'baz' on 'Foo'
```
- Also note that we need a special construct to keep track of whether a
constructor enum variant has been invoked or not. In the previous
implementation, we'd know this because the variant's type was a `Type::Fn`
whose return type was the `Type::Enum`. Now, we use a slightly less
ideal `is_constructed` value on the `Type::EnumVariant` type. Kinda meh,
but it works.