Match (#208)
* Working towards #191 (#192)

- Add lexing and parsing for match statements/expressions, handling type
cases with optional re-binding variables.

* Typechecking match statements/exressions (#198)

* Typechecking match statements/exressions

* Removing usages of Vec::remove_item, since it's unstable

* Match  compiling (#202)

* Typechecking match statements/exressions

* Removing usages of Vec::remove_item, since it's unstable

* Compiling match statements/expressions

- Compiling cases of None and _

* Match statements/expressions: handling Type cases

- Match statements and expressions can now match a value against a given
type! So the things that can be matched against now include `None`,
types, and a wildcard (`_`).
- This also refactors `Prelude` to use `thread_local!` for static global
constants, to help tidy things up a bit.
- This also adds a way to compare types for equality (as is necessary in
order to facilitate Type branches in `match` statements/expressions);
right now it just compares the name, but it should probably be more
robust than that in the future.

* Adding typechecker error for unimplemented feature

* Match expressions: enum cases (#204)

* Match expressions: enum cases

- Add ability to match on enums in match statements/expressions.
This does not currently support destructuring
(ie.  `Color.RGB(r, g, b) => ...`), but that's coming soon

* Adding enum variants as separate types

- Add new type `Type::EnumVariant`, which is a nicer abstraction over
top of enum variants, allowing each enum's variant to represent a
separate discrete type. This allows us to access fields of an enum
variant, ie
```
enum Foo { Bar(baz: Int) }
val f = Foo.Bar(baz: 24)
f.baz
```
Since the typechecker now knows that the type of `f` is not only `Foo`,
but `Foo.Bar`, it allows us to access `.baz` off of it. In prior
implementations, this would have been typechecked as:
```
val f: Foo = Foo.Bar(baz: 24)
f.baz
 ^ Error, no member called 'baz' on 'Foo'
```
- Also note that we need a special construct to keep track of whether a
constructor enum variant has been invoked or not. In the previous
implementation, we'd know this because the variant's type was a `Type::Fn`
whose return type was the `Type::Enum`. Now, we use a slightly less
ideal `is_constructed` value on the `Type::EnumVariant` type. Kinda meh,
but it works.

* Match statements: enum destructuring (#207)

- Allow for the destructuring of enum variants in match
statements/expressions. As of right now, this only applies to enum
variants; attempting to destructure other types will present an error.