Maps gets a huge overhaul
- Maps no longer implicitly convert to Structs if all the fields match
up
- Remove notion of "homogeneously-typed", now we just track the key_type
and the value_type. It's much more sane this way, and much more in line
with how people expect to interact with maps.
- Maps can now be `Map<Value, Value>`, meaning that the keys can be
_any_ value, not just String values! It's currently not possible to
create a map literal with arbitrary expressions as keys (see #218),
however.
- Types (such as `Array` and `Map`) are now available in word-form in
type annotations. This was done in an admittedly sub-optimal way, by
utilizing the idea of referenceable_types. It might have to change in
the future but honestly it's not too bad and makes everything else
pretty easy. The only thing that's a _little_ weird is that, when
resolving referenceable_types, if its underlying type is a struct/enum,
we return the `Type::Reference`; if it's an array or a map, we hydrate
the reference and return the `Type::Array`/`Type::Map`. _Little_ weird,
but still pretty sane.
- Add NativeType scaffolding for Maps as well, with some methods made
available.