Fixing recursion bug
- There was a bug where non-root-level recursive functions which
recursed within a lambda were not correctly tagged as recursive calls,
so the functions were unknown at compile-time. This was fixed in the
typechecker.
- There was _another_ bug where, if code within a lambda (which gets
executed from a native context) does not return a successful `Result`,
that error will be swallowed and the function will instead return
`Value::Nil`. This was super confusing! Now, if such a lambda encounters
an error, the process will halt and the error will be printed as a
runtime error. There is currently no way to recover from runtime errors
(but they should also be impossible, aside from a language
implementation bug).
- Also increased the max stack frame count from 64 to 1024.