Fixing issue with if-exprs
- There was a problem with if-exprs that were missing an else-block;
since each branch didn't produce a value, assignment of the result of
the if-expr when the main block is _not_ entered is missing. So what
happens at runtime is the top value of the stack ends up being used
instead (whatever random value that happens to be). To mitigate this,
let's modify the AST while typechecking to inject a `None` ident into
would-be missing else-blocks. (Note that this means we need to do an
_extra_ cleanup step when treating ending if-stmts in functions as
exprs - if we had injected a `None` into the else block, we need to
un-inject that. There should _definitely_ be a better way of handling
this...)
- Made change to type equality, `Unit` should not be assignable to
`Any`