Breaking change to function signatures
- Previously, functions' return types could be inferred from the body.
However, this inference became very complicated, especially when trying
to typecheck functions which reference not-yet-defined functions. So for
now, the breaking change is that **all functions will now be required to
have a return type annotation**. This could change in the future (I
hope), but it certainly makes a lot of things easier for now.
- Because of this new requirement, we can actually clean a bunch of
stuff up. Return types can no longer by `Unknown`, so we get to remove a
bunch of code related to that. All functions will have return types, so
we don't need to handle recursive functions separately; so that error
can be deleted entirely. Same goes for the `MissingRequiredReturnType`
error that had previously been thrown for methods within a typedef. Good
cleanup all around!