Fixing minor bug
- The prior implementation of "hoisting" allowed for potential
situations where functions were accessed before they were actually
defined. This can lead to all sorts of weird scenarios, like:
```
if true {
  if true {
    abc() // This references `a`, which hasn't been defined yet
  }
  var a = "hello"
  func abc() = println(a)
}
```
To mitigate this situation, we keep track of the function definitions in
each scope in a first pass over that scope, and then use those
definitions as part of binding resolution. If we've crossed a function
boundary, then fns in outer scopes are eligible for calling.