Functions can be referened out-of-order
- Functions in a given scope can now be referenced and called out of
declaration order. This applies to the root-level scope, nested function
scopes, if-/match-expressions, if while-/for-loops.
- This also fixes an upvalue issue that had been present before. The
following code would not work before:
```
func a() {
  func b() = println("hello")
  func c() {
    func d() = b()
  }
}
```
Since `b` was not in `c`'s scope, we have to look up an additional level
to `a`'s scope; it's there, so we create an upvalue in `a`'s scope and
also in `c`'s scope. This was fine, this had worked before. The problem
however was in the code to peel upvalues off of the array based on
fn_depth. Since the intermediate, transitive upvalue was added in
between 2 higher-depth upvalues, the existing code would not pick it off
of the list. The updated logic now fixes that.