Destructuring: Tuples
- Refactoring to make the parsing/typechecking/compilation more
streamlined and reusable. Also, nested destructuring is now supported for
tuples!
- This also changes the `val`/`var` paradigm a _tiny bit_; for
destructured values, it is now impossible to have uninitialized
variables (eg, `var (a, b, c)`). This will actually fail to _parse_,
since the parser expects an `=` token after the closing `)`. This is a
good mental model I think (not to mention that it cleans up the compiler
code a good bit), since the argument here is that we're not really
destructuring anything. We're just declaring multiple variables; why not
just do
```
var a: Int
var b: Int
var c: Int
```