Varargs: typechecking
- Change FnType to have an `is_variadic` field, which is true if
one of the parameters in the function's type is variadic. This involved
touching a lot of the autogenerated code, which was a bit annoying, but
it all worked out.
- When typechecking invocations, if the last parameter is a variadic
parameter, collect the remaining passed parameters and typecheck them
against the variadic parameter's type. The emitted typed ast node will
be a transformed value (so no compilation work need be done). If there
are no parameters passed, the default value (of []) will be used.