Begin implementing binding API for native types
- The old native types method was interesting but it had lots of
drawbacks. For one, it didn't support having fields in types, without
having a special implementation of `Obj`. This re-imagining does a lot
of things better:
  - Uses a proc-macro to abstract a lot of the messiness away, rather
than auto-generate an absolutely massive rust file, which would only
ever get bigger. This macro provides a super nice API to bind a rust
struct to an abra type, and a much much nicer way of dealing with
arguments. It also allows for getters and setters, and for constructor
bindings.
  - It abstracts everything behind `Obj::NativeInstanceObj` so we will
eventually only have a few things grouped under `Obj`.