Lazy instance method binding
- Eliminate the `Init` opcode; now, when an instance is constructed,
only the `New` opcode is used which attaches the `TypeValue` to a given
instance. Methods are now no longer stored directy on instances, but
rather are cloned from the `TypeValue` and bound to the instance upon
`GetMethod`. This cleans things up and also reduces the size of each
instance, since they won't each have duplicates of every method (one
bound, one yet-unbound).