Migrating NativeInt and NativeFloat
- Amend the binding framework to allow for "pseudomethods" and
"pseudotypes", which facilitates binding methods onto NativeInt and
NativeFloat. Basically, these allow for the old way that NativeTypes
were handled, which was basically a collection of NativeFn instances
that were bound at GetMethod-time to the Value::Int/Value::Float. This
emulates that old behavior.
  - I also added a "noconstruct" parameter to #[abra_type], allowing for
the omission of a function bound with #[abra_constructor]. This is nice
for the NativeInt/NativeFloat use case, but also for NativeString. I
don't expect this to be useful in other future types, since other types
will be constructed via normal constructor functions and will not have
language-level syntax to act as a constructor (there may be a few
exceptions of course).