Improving performance for native types
- After using AOC2020 Day 22 as a benchmark, I noticed that the
performance degraded a lot (~20s to ~95s) after the native types
refactor. So I needed to tidy things up a bit, adding native
implementations for String, Array, and Set. The ultimate goal here is
ultimately to retire the `Value::Obj` variant (and thus, the `Obj`
enum entirely), in favor of more top-level `Value` variants. These will
be easier to work with as it won't involve as much ambiguity when
borrowing the `Arc`s. This will become more relevant when implementing
gc.
- This also removes the `derive(PartialOrd)` from all `Value` variant
substructs, since it's really unnecessary.